+++ src/ExAs/ComparableAssertions.cs
@@ -12,10 +12,10 @@ namespace ExAs
             return member.SetAssertion(new IsSmallerAssertion<TMember>(expected));
         }
 
-        public static IAssert<T> IsBiggerThan<T, TMember>(this IAssertMember<T, TMember> member, TMember expected)
+        public static IAssert<T> IsGreaterThan<T, TMember>(this IAssertMember<T, TMember> member, TMember expected)
             where TMember : IComparable<TMember>
         {
-            return member.SetAssertion(new IsBiggerAssertion<TMember>(expected));
+            return member.SetAssertion(new GreaterThanAssertion<TMember>(expected));
         }
 
         public static IAssert<T> IsInRange<T, TMember>(this IAssertMember<T, TMember> member, TMember min, TMember max)