+++ src/ExAs/ShortAssertions.cs
@@ -16,9 +16,9 @@ namespace ExAs
             return member.SetAssertion(new IsSmallerAssertion<short>(expected));
         }
 
-        public static IAssert<T> IsBiggerThan<T>(this IAssertMember<T, short> member, short expected)
+        public static IAssert<T> IsGreaterThan<T>(this IAssertMember<T, short> member, short expected)
         {
-            return member.SetAssertion(new IsBiggerAssertion<short>(expected));
+            return member.SetAssertion(new GreaterThanAssertion<short>(expected));
         }
 
         public static IAssert<T> IsInRange<T>(this IAssertMember<T, short> member, short min, short max)