+++ tests/ExAs.Tests/Api/DoubleAssertionFeature.cs
@@ -53,10 +53,10 @@ namespace ExAs.Api
         }
 
         [Test]
-        public void IsBiggerThan_Expected37_8_Get38_8_ShouldPass()
+        public void IsGreaterThan_Expected37_8_Get38_8_ShouldPass()
         {
             // Act
-            var result = padavanNaruto.Evaluate(n => n.Member(x => x.SkillValue).IsBiggerThan(37.8));
+            var result = padavanNaruto.Evaluate(n => n.Member(x => x.SkillValue).IsGreaterThan(37.8));
 
             // Assert
             result.ExAssert(r => r.IsNotNull()
@@ -65,10 +65,10 @@ namespace ExAs.Api
         }
 
         [Test]
-        public void IsBiggerThan_Expected100_1_Get99_7_ShouldFail()
+        public void IsGreaterThan_Expected100_1_Get99_7_ShouldFail()
         {
             // Act
-            var result = skilledNaruto.Evaluate(n => n.Member(x => x.SkillValue).IsBiggerThan(100.1));
+            var result = skilledNaruto.Evaluate(n => n.Member(x => x.SkillValue).IsGreaterThan(100.1));
 
             // Assert
             result.ExAssert(r => r.IsNotNull()