+++ tests/ExAs.Tests/Api/ShortAssertionFeature.cs
@@ -29,10 +29,10 @@ namespace ExAs.Api
         }
 
         [Test]
-        public void IsBiggerThan_ExpectingBiggerThan64_OnNinjaWith65_ShouldSucceed()
+        public void IsGreaterThan_ExpectingBiggerThan64_OnNinjaWith65_ShouldSucceed()
         {
             // act
-            var result = strongNinja.Evaluate(n => n.Member(x => x.strength).IsBiggerThan(64));
+            var result = strongNinja.Evaluate(n => n.Member(x => x.strength).IsGreaterThan(64));
 
             // assert
             result.ExAssert(r => r.Fullfills(true, "DetailedNinja: ( )strength = 65", "(expected: bigger than 64)"));